refactored restaurant code 450 lines now, functionality is same
