+++ lib/screens/dashboard/restaurant/dialogs/notifications_dialog.dart
@@ -1,0 +1,26 @@
+import 'package:flutter/material.dart';
+
+class NotificationsDialog extends StatelessWidget {
+  const NotificationsDialog({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return AlertDialog(
+      title: const Text('Notifications'),
+      content: const Column(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          Text('• New donation available nearby'),
+          Text('• Claim status updated'),
+          Text('• Monthly impact report ready'),
+        ],
+      ),
+      actions: [
+        TextButton(
+          onPressed: () => Navigator.of(context).pop(),
+          child: const Text('Close'),
+        ),
+      ],
+    );
+  }
+}
