+++ lib/screens/dashboard/restaurant/tabs/recipes/widgets/multi_ingredient_recipe_generator.dart
@@ -14,7 +14,7 @@ class MultiIngredientRecipeGenerator extends StatelessWidget {
   Widget build(BuildContext context) {
     return Container(
       width: double.infinity,
-      padding: const EdgeInsets.all(24),
+      padding: const EdgeInsets.all(12),
       decoration: BoxDecoration(
         color: Colors.white,
         borderRadius: BorderRadius.circular(16),
@@ -27,70 +27,72 @@ class MultiIngredientRecipeGenerator extends StatelessWidget {
           ),
         ],
       ),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Row(
-            children: [
-              Container(
-                padding: const EdgeInsets.all(8),
-                decoration: BoxDecoration(
-                  color: Colors.green.withOpacity(0.1),
-                  borderRadius: BorderRadius.circular(8),
+      child: SingleChildScrollView(
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            Row(
+              children: [
+                Container(
+                  padding: const EdgeInsets.all(8),
+                  decoration: BoxDecoration(
+                    color: Colors.green.withOpacity(0.1),
+                    borderRadius: BorderRadius.circular(8),
+                  ),
+                  child: const Icon(
+                    Icons.auto_awesome,
+                    color: Colors.green,
+                    size: 20,
+                  ),
                 ),
-                child: const Icon(
-                  Icons.auto_awesome,
-                  color: Colors.green,
-                  size: 20,
+                const SizedBox(width: 12),
+                const Text(
+                  'Multi-Ingredient Recipe Generator',
+                  style: TextStyle(
+                    fontSize: 18,
+                    fontWeight: FontWeight.bold,
+                    color: Colors.black87,
+                  ),
+                ),
+              ],
+            ),
+            const SizedBox(height: 12),
+            const Text(
+              'Generate complex recipes using multiple ingredients from your inventory:',
+              style: TextStyle(
+                color: Colors.grey,
+                fontSize: 13,
+              ),
+            ),
+            const SizedBox(height: 12),
+            SizedBox(
+              width: double.infinity,
+              child: ElevatedButton.icon(
+                onPressed: viewModel.inventory.length >= 2 ? () => onGenerateAdvancedRecipe() : null,
+                icon: const Icon(Icons.auto_awesome),
+                label: const Text('Generate Multi-Ingredient Recipe'),
+                style: ElevatedButton.styleFrom(
+                  backgroundColor: Colors.green,
+                  foregroundColor: Colors.white,
+                  padding: const EdgeInsets.symmetric(vertical: 12),
+                  shape: RoundedRectangleBorder(
+                    borderRadius: BorderRadius.circular(12),
+                  ),
                 ),
               ),
-              const SizedBox(width: 12),
-              const Text(
-                'Multi-Ingredient Recipe Generator',
+            ),
+            if (viewModel.inventory.length < 2) ...[
+              const SizedBox(height: 8),
+              Text(
+                'Add at least 2 items to your inventory to use this feature',
                 style: TextStyle(
-                  fontSize: 18,
-                  fontWeight: FontWeight.bold,
-                  color: Colors.black87,
+                  color: Colors.grey[500],
+                  fontSize: 11,
                 ),
               ),
             ],
-          ),
-          const SizedBox(height: 16),
-          const Text(
-            'Generate complex recipes using multiple ingredients from your inventory:',
-            style: TextStyle(
-              color: Colors.grey,
-              fontSize: 14,
-            ),
-          ),
-          const SizedBox(height: 20),
-          SizedBox(
-            width: double.infinity,
-            child: ElevatedButton.icon(
-              onPressed: viewModel.inventory.length >= 2 ? () => onGenerateAdvancedRecipe() : null,
-              icon: const Icon(Icons.auto_awesome),
-              label: const Text('Generate Multi-Ingredient Recipe'),
-              style: ElevatedButton.styleFrom(
-                backgroundColor: Colors.green,
-                foregroundColor: Colors.white,
-                padding: const EdgeInsets.symmetric(vertical: 16),
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(12),
-                ),
-              ),
-            ),
-          ),
-          if (viewModel.inventory.length < 2) ...[
-            const SizedBox(height: 12),
-            Text(
-              'Add at least 2 items to your inventory to use this feature',
-              style: TextStyle(
-                color: Colors.grey[500],
-                fontSize: 12,
-              ),
-            ),
           ],
-        ],
+        ),
       ),
     );
   }
